feat: implement loan store on index.tsx
